Resolve wear todos
