+++ wear/src/main/java/ru/beryukhov/coffeegram/wear/WearActivity.kt
@@ -3,10 +3,9 @@ package ru.beryukhov.coffeegram.wear
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
@@ -28,6 +27,7 @@ import androidx.wear.compose.material.Text
 import kotlinx.coroutines.flow.MutableStateFlow
 import ru.beryukhov.coffeegram.data.CoffeeType
 import ru.beryukhov.coffeegram.data.DayCoffee
+import ru.beryukhov.coffeegram.data.withEmpty
 
 val coffeeState: MutableStateFlow<DayCoffee> by lazy { MutableStateFlow(value = DayCoffee()) }
 
@@ -47,7 +47,7 @@ class WearActivity : ComponentActivity() {
 internal fun PagesContent() {
     val dayCoffee by coffeeState.collectAsState()
     ScalingLazyColumn(
-        modifier = Modifier.fillMaxHeight(),
+        modifier = Modifier.fillMaxSize(),
         contentPadding = PaddingValues(
             top = 28.dp,
             start = 10.dp,
@@ -86,18 +86,4 @@ fun CoffeeItem(c: CoffeeType, count: Int, modifier: Modifier = Modifier) {
         },
         onClick = {},
     )
-}
-
-// todo share it
-@VisibleForTesting
-internal fun Map<CoffeeType, Int>.withEmpty(): List<Pair<CoffeeType, Int>> {
-    @Suppress("DataClassShouldBeImmutable")
-    data class MutablePair(val ct: CoffeeType, var count: Int)
-
-    val emptyList: MutableList<MutablePair> =
-        CoffeeType.entries.map { MutablePair(it, 0) }.toMutableList()
-    this.forEach { entry: Map.Entry<CoffeeType, Int> ->
-        emptyList.filter { it.ct == entry.key }.forEach { it.count = entry.value }
-    }
-    return emptyList.map { it.ct to it.count }
 }
